refactor: RxBaseTableCell CommonUI 모듈로 이동
